+++ tests/test_payu.py
@@ -1354,7 +1354,7 @@ class TestPayuProvider(TestCase):
             amount = self.provider.refund(self.payment, Decimal(110))
 
         self.assertEqual(refund_request_mock.call_count, 1)
-        self.assertEqual(amount, Decimal(110))
+        self.assertEqual(amount, Decimal(0))
         self.assertEqual(self.payment.total, Decimal(220))
         self.assertEqual(self.payment.captured_amount, Decimal(210))
         self.assertEqual(self.payment.status, PaymentStatus.CONFIRMED)
@@ -1417,7 +1417,7 @@ class TestPayuProvider(TestCase):
             amount = self.provider.refund(self.payment)
 
         self.assertEqual(refund_request_mock.call_count, 1)
-        self.assertEqual(amount, Decimal(220))
+        self.assertEqual(amount, Decimal(0))
         self.assertEqual(self.payment.total, Decimal(220))
         self.assertEqual(self.payment.captured_amount, Decimal(220))
         self.assertEqual(self.payment.status, PaymentStatus.CONFIRMED)
@@ -1482,7 +1482,7 @@ class TestPayuProvider(TestCase):
                 amount = self.provider.refund(self.payment, Decimal(110))
 
         self.assertEqual(refund_request_mock.call_count, 1)
-        self.assertEqual(amount, Decimal(110))
+        self.assertEqual(amount, Decimal(0))
         self.assertEqual(self.payment.total, Decimal(220))
         self.assertEqual(self.payment.captured_amount, Decimal(220))
         self.assertEqual(self.payment.status, PaymentStatus.CONFIRMED)
@@ -1544,7 +1544,7 @@ class TestPayuProvider(TestCase):
             amount = self.provider.refund(self.payment, Decimal(110))
 
         self.assertEqual(refund_request_mock.call_count, 1)
-        self.assertEqual(amount, Decimal(110))
+        self.assertEqual(amount, Decimal(0))
         self.assertEqual(self.payment.total, Decimal(220))
         self.assertEqual(self.payment.captured_amount, Decimal(220))
         self.assertEqual(self.payment.status, PaymentStatus.CONFIRMED)
@@ -1608,7 +1608,7 @@ class TestPayuProvider(TestCase):
 
         self.assertEqual(refund_request_mock.call_count, 2)
         self.assertEqual(amount2, amount1)
-        self.assertEqual(amount2, Decimal(200))
+        self.assertEqual(amount2, Decimal(0))
         self.assertEqual(self.payment.total, Decimal(220))
         self.assertEqual(self.payment.captured_amount, Decimal(220))
         self.assertEqual(self.payment.status, PaymentStatus.CONFIRMED)
@@ -1626,7 +1626,7 @@ class TestPayuProvider(TestCase):
         )
         self.assertFalse(caught_warnings)
 
-    def test_refund_finalized(self):
+    def test_refund_pending(self):
         with warnings.catch_warnings(record=True) as caught_warnings:
             warnings.simplefilter("always")
             self.set_up_provider(
@@ -1648,7 +1648,7 @@ class TestPayuProvider(TestCase):
                 "currencyCode": "USD",
                 "description": "desc 1234 110",
                 "creationDateTime": "2020-07-02T09:19:03.896+02:00",
-                "status": "FINALIZED",
+                "status": "PENDING",
                 "statusDateTime": "2020-07-02T09:19:04.013+02:00",
             },
             "status": {
@@ -1671,7 +1671,7 @@ class TestPayuProvider(TestCase):
             amount = self.provider.refund(self.payment, Decimal(110))
 
         self.assertEqual(refund_request_mock.call_count, 1)
-        self.assertEqual(amount, Decimal(110))
+        self.assertEqual(amount, Decimal(0))
         self.assertEqual(self.payment.total, Decimal(220))
         self.assertEqual(self.payment.captured_amount, Decimal(220))
         self.assertEqual(self.payment.status, PaymentStatus.CONFIRMED)